refactor: modularise stack header (#150)
This change introduces the code and frame header files with content
taken out of the stack header for better source code modularity.
